phrases/requirements.txt
pdftotext
textblob
python-magic
